IDEA-144052 "path.macros.xml" only
(cherry picked from commit b51aca4f1b8ea569dbe4021c9f5dc252da62c07a)

+++ platform/configuration-store-impl/src/ApplicationStoreImpl.kt
@@ -87,6 +87,10 @@ class ApplicationStorageManager(application: Application, pathMacroManager: Path
 
   override fun dataLoadedFromProvider(storage: FileBasedStorage, element: Element?) {
     // IDEA-144052 When "Settings repository" is enabled changes in 'Path Variables' aren't saved to default path.macros.xml file causing errors in build process
+    if (storage.fileSpec != "path.macros.xml") {
+      return
+    }
+
     LOG.catchAndLog {
       if (element == null) {
         storage.file.delete()
